Finalize the details
