--- src/pages/OpenDevMonthlyCalls/index.mdx
@@ -1,11 +1,0 @@
-The Polkadot Fellowship runs a call (“OpenDev”) on a monthly basis to share current and future developments of the Polkadot roadmap.
-
-## Agenda
-
-All new and existing members are invited to join this live call to introduce themselves, present their contributions to the codebase, participate in discussions about RFCs, propose ideas for the growth of the Fellowship, and answer questions from the general public.
-
-## Upcoming call
-
-#### The next monthly call is scheduled for: **To Be Announced**
-
-## Past calls
